+++ src/main/scala/li/cil/oc/client/gui/Screen.scala
@@ -111,10 +111,10 @@ class Screen(val buffer: api.internal.TextBuffer, val hasMouse: Boolean, val has
   }
 
   override def drawBuffer(stack: PoseStack) = {
-    stack.translate(x, y, 0)
+    stack.translate(x.toFloat, y.toFloat, 0f)
     BufferRenderer.drawBackground(stack, innerWidth, innerHeight)
     if (hasPower()) {
-      stack.translate(bufferMargin, bufferMargin, 0)
+      stack.translate(bufferMargin.toFloat, bufferMargin.toFloat, 0f)
       stack.scale(scale.toFloat, scale.toFloat, 1)
       BufferRenderer.drawText(stack, buffer)
     }
