+++ src/main/scala/li/cil/oc/client/gui/traits/InputBuffer.scala
@@ -95,10 +95,10 @@ trait InputBuffer extends DisplayBuffer {
 
       val t = Tesselator.getInstance
       val r = t.begin(VertexFormat.Mode.QUADS, DefaultVertexFormat.POSITION_TEX)
-      r.addVertex(stack.last.pose, x, y + 16, 0).setUv(0, 1)
-      r.addVertex(stack.last.pose, x + 16, y + 16, 0).setUv(1, 1)
-      r.addVertex(stack.last.pose, x + 16, y, 0).setUv(1, 0)
-      r.addVertex(stack.last.pose, x, y, 0).setUv(0, 0)
+      r.addVertex(stack.last.pose, x.toFloat, (y + 16).toFloat, 0f).setUv(0, 1)
+      r.addVertex(stack.last.pose, (x + 16).toFloat, (y + 16).toFloat, 0f).setUv(1, 1)
+      r.addVertex(stack.last.pose, (x + 16).toFloat, y.toFloat, 0f).setUv(1, 0)
+      r.addVertex(stack.last.pose, x.toFloat, y.toFloat, 0f).setUv(0, 0)
       BufferUploader.drawWithShader(r.buildOrThrow())
 
       RenderState.checkError(getClass.getName + ".drawBufferLayer: keyboard icon")
